requirements.txt
django
djangorestframework
django-filter
django-environ
drf-spectacular
requests
celery
redis
psycopg2-binary
python-dotenv